Move stuff to provider packages
