+++ airflow/io/__init__.py
@@ -52,22 +52,10 @@ S3_ACCESS_KEY_ID = "s3.access-key-id"
 S3_SECRET_ACCESS_KEY = "s3.secret-access-key"
 S3_SESSION_TOKEN = "s3.session-token"
 S3_REGION = "s3.region"
-S3_PROXY_URI = "s3.proxy-uri"
 HDFS_HOST = "hdfs.host"
 HDFS_PORT = "hdfs.port"
 HDFS_USER = "hdfs.user"
 HDFS_KERB_TICKET = "hdfs.kerberos_ticket"
-GCS_TOKEN = "gcs.oauth2.token"
-GCS_TOKEN_EXPIRES_AT_MS = "gcs.oauth2.token-expires-at"
-GCS_PROJECT_ID = "gcs.project-id"
-GCS_ACCESS = "gcs.access"
-GCS_CONSISTENCY = "gcs.consistency"
-GCS_CACHE_TIMEOUT = "gcs.cache-timeout"
-GCS_REQUESTER_PAYS = "gcs.requester-pays"
-GCS_SESSION_KWARGS = "gcs.session-kwargs"
-GCS_ENDPOINT = "gcs.endpoint"
-GCS_DEFAULT_LOCATION = "gcs.default-bucket-location"
-GCS_VERSION_AWARE = "gcs.version-aware"
 
 TOKEN = "token"
