+++ airflow/io/fsspec.py
@@ -18,180 +18,42 @@
 import errno
 import logging
 import os
-from functools import lru_cache, partial
+from functools import lru_cache
 from typing import (
-    Any,
     Callable,
-    Dict,
     Union,
 )
 from urllib.parse import urlparse
 
-import requests
-from botocore import UNSIGNED
-from botocore.awsrequest import AWSRequest
 from fsspec import AbstractFileSystem
 from fsspec.implementations.local import LocalFileSystem
-from requests import HTTPError
 
-from airflow.hooks.base import BaseHook
-
+import airflow.providers
 from airflow.io import (
-    GCS_ACCESS,
-    GCS_CACHE_TIMEOUT,
-    GCS_CONSISTENCY,
-    GCS_DEFAULT_LOCATION,
-    GCS_ENDPOINT,
-    GCS_REQUESTER_PAYS,
-    GCS_SESSION_KWARGS,
-    GCS_VERSION_AWARE,
-    S3_PROXY_URI,
     FileIO,
     InputFile,
     InputStream,
     OutputFile,
     OutputStream,
-    Properties,
-    TOKEN,
 )
-from airflow.io.exceptions import SignError
-from airflow.providers.amazon.aws.hooks.base_aws import AwsGenericHook
-from airflow.providers.google.common.hooks.base_google import GoogleBaseHook
-from airflow.providers.microsoft.azure.utils import get_field
+from airflow.stats import Stats
+from airflow.utils.module_loading import iter_namespace
 
 log = logging.getLogger(__name__)
-
-
-def s3v4_rest_signer(properties: Properties, request: AWSRequest, **_: Any) -> AWSRequest:
-    if TOKEN not in properties:
-        raise SignError("Signer set, but token is not available")
-
-    signer_url = properties["uri"].rstrip("/")
-    signer_headers = {"Authorization": f"Bearer {properties[TOKEN]}"}
-    signer_body = {
-        "method": request.method,
-        "region": request.context["client_region"],
-        "uri": request.url,
-        "headers": {key: [val] for key, val in request.headers.items()},
-    }
-
-    response = requests.post(f"{signer_url}/v1/aws/s3/sign", headers=signer_headers, json=signer_body)
-    try:
-        response.raise_for_status()
-        response_json = response.json()
-    except HTTPError as e:
-        raise SignError(f"Failed to sign request {response.status_code}: {signer_body}") from e
-
-    for key, value in response_json["headers"].items():
-        request.headers.add_header(key, ", ".join(value))
-
-    request.url = response_json["uri"]
-
-    return request
-
-
-SIGNERS: Dict[str, Callable[[Properties, AWSRequest], AWSRequest]] = {"S3V4RestSigner": s3v4_rest_signer}
-
 
 def _file(_: str | None) -> LocalFileSystem:
     return LocalFileSystem()
 
 
-def _s3(conn_id: str | None) -> AbstractFileSystem:
-    aws = AwsGenericHook(aws_conn_id=conn_id)
-    from s3fs import S3FileSystem
-
-    client_kwargs = {
-        "endpoint_url": aws.conn_config.endpoint_url,
-        "aws_access_key_id": aws.conn_config.aws_access_key_id,
-        "aws_secret_access_key": aws.conn_config.aws_secret_access_key,
-        "aws_session_token": aws.conn_config.aws_session_token,
-        "region_name": aws.conn_config.region_name,
-    }
-    config_kwargs = {}
-    register_events: Dict[str, Callable[[Properties], None]] = {}
-
-    if signer := aws.conn_config.extra_config.get("s3.signer"):
-        log.info("Loading signer %s", signer)
-        if singer_func := SIGNERS.get(signer):
-            properties: Properties = {
-                "uri": aws.conn_config.extra_config.get("s3.signer_uri"),
-                TOKEN: aws.conn_config.extra_config.get("s3.signer_token"),
-            }
-            singer_func_with_properties = partial(singer_func, properties)
-            register_events["before-sign.s3"] = singer_func_with_properties
-
-            # Disable the AWS Signer
-            config_kwargs["signature_version"] = UNSIGNED
-        else:
-            raise ValueError(f"Signer not available: {signer}")
-
-    if proxy_uri := aws.conn_config.extra_config.get(S3_PROXY_URI):
-        config_kwargs["proxies"] = {"http": proxy_uri, "https": proxy_uri}
-
-    fs = S3FileSystem(client_kwargs=client_kwargs, config_kwargs=config_kwargs)
-
-    for event_name, event_function in register_events.items():
-        fs.s3.meta.events.register_last(event_name, event_function, unique_id=1925)
-
-    return fs
-
-
-def _gs(conn_id: str | None) -> AbstractFileSystem:
-    # https://gcsfs.readthedocs.io/en/latest/api.html#gcsfs.core.GCSFileSystem
-    from gcsfs import GCSFileSystem
-
-    g = GoogleBaseHook(gcp_conn_id=conn_id)
-    creds = g.get_credentials()
-
-    return GCSFileSystem(
-        project=g.project_id,
-        access=g.extras.get(GCS_ACCESS, "full_control"),
-        token=creds.token,
-        consistency=g.extras.get(GCS_CONSISTENCY, "none"),
-        cache_timeout=g.extras.get(GCS_CACHE_TIMEOUT),
-        requester_pays=g.extras.get(GCS_REQUESTER_PAYS, False),
-        session_kwargs=g.extras.get(GCS_SESSION_KWARGS, {}),
-        endpoint_url=g.extras.get(GCS_ENDPOINT),
-        default_location=g.extras.get(GCS_DEFAULT_LOCATION),
-        version_aware=g.extras.get(GCS_VERSION_AWARE, "false").lower() == "true",
-    )
-
-
-def _adlfs(conn_id: str | None) -> AbstractFileSystem:
-    from adlfs import AzureBlobFileSystem
-
-    conn = BaseHook.get_connection(conn_id)
-    extras = conn.extra_dejson
-
-    connection_string = get_field(conn_id=conn_id, conn_type="azure_data_lake", extras=extras, field_name="connection_string")
-    account_name = get_field(conn_id=conn_id, conn_type="azure_data_lake", extras=extras, field_name="account_name")
-    account_key = get_field(conn_id=conn_id, conn_type="azure_data_lake", extras=extras, field_name="account_key")
-    sas_token = get_field(conn_id=conn_id, conn_type="azure_data_lake", extras=extras, field_name="sas_token")
-    tenant = get_field(conn_id=conn_id, conn_type="azure_data_lake", extras=extras, field_name="tenant")
-
-    return AzureBlobFileSystem(
-        connection_string=connection_string,
-        account_name=account_name,
-        account_key=account_key,
-        sas_token=sas_token,
-        tenant_id=tenant,
-        client_id=conn.login,
-        client_secret=conn.password,
-    )
-
-
+# always support local file systems
 SCHEME_TO_FS = {
     "file": _file,
-    "s3": _s3,
-    "s3a": _s3,
-    "s3n": _s3,
-    "abfs": _adlfs,
-    "abfss": _adlfs,
-    "adl": _adlfs,
-    "gs": _gs,
-    "gcs": _gs,
 }
+
+
+def _register_schemes() -> None:
+    with Stats.timer("airflow.io.load_filesystems") as timer:
+        for _, name, _ in iter_namespace(airflow.providers):
 
 
 class FsspecInputFile(InputFile):
